requirements.txt
alembic==1.13.1
APScheduler==3.10.4
async-timeout==4.0.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==1.16.0
eventlet==0.33.3
Flask==2.3.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.3.2
MarkupSafe==2.1.3
packaging==23.2
pika==1.3.2
pdfminer.six==20231228
psycopg2-binary==2.9.9
pymongo==4.6.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==2.3.6
zipp==3.17.0
nltk==3.8.1